requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
Pillow~=9.4.0
accounts~=0.2.4